Make MiniInterp own WaveSignalTrace
