+++ monitor/src/main.rs
@@ -9,7 +9,7 @@ mod signal_trace;
 
 use crate::designs::{Instance, collects_design_names, find_designs, parse_instance};
 use crate::mini_interp::MiniInterpreter;
-use crate::signal_trace::{PortKey, SignalTrace, WaveSamplingMode, WaveSignalTrace};
+use crate::signal_trace::{WaveSamplingMode, WaveSignalTrace};
 use clap::Parser;
 use clap_verbosity_flag::{Verbosity, WarnLevel};
 use protocols::diagnostic::DiagnosticHandler;
@@ -87,35 +87,7 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
     let (transaction, symbol_table) = &transactions_symbol_tables[0];
 
     // Create a new Interpreter for the `.prot` file
-    let mut interpreter = MiniInterpreter::new(transaction, symbol_table, &trace, design);
-
-    for port_key in trace.port_map.keys() {
-        // We assume that there is only one `Instance` at the moment
-        let PortKey {
-            instance_id,
-            pin_id,
-        } = port_key;
-
-        // Fetch the current value of the `pin_id`
-        // (along with the name of the corresponding `Field`)
-        let current_value = trace.get(*instance_id, *pin_id);
-        let field_name = design
-            .get_pin_name(pin_id)
-            .unwrap_or_else(|| panic!("Missing pin_id {} in design.pins", pin_id));
-        println!(
-            "{} ({}) |-> {:#?} ",
-            field_name,
-            pin_id,
-            current_value.clone()
-        );
-
-        // TODO: figure out what to do if the `pin_id` already has a value in the environment
-
-        // TODO: figure out how to handle `step()`
-
-        // Update the `args_mapping` with the `current_value` for the `pin_id`
-        interpreter.update_arg_value(*pin_id, current_value);
-    }
+    let interpreter = MiniInterpreter::new(transaction, symbol_table, trace, design);
 
     Ok(())
 }
